pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "THEBKGOSWAMI08",email = "goswami.bk555@gmail.com"}
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "groq (<1.0.0)",
    "langchain-core (>=1.3.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.1.2,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "sentence-transformers (>=5.4.1,<6.0.0)",
    "pypdf (>=6.10.2,<7.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "streamlit (>=1.56.0,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
